ルームの編集ロジックをカスタムフック化
